# src/main/resources/static/openapi.yaml
openapi: 3.0.3
info:
  title: Simple Mall API
  description: Spring Boot mall API with MySQL, session login, filters, Swagger, and optional Kafka events.
  version: v3.0
  contact:
    name: Carlosjjh
  license:
    name: MIT-style demo project
servers:
  - url: http://localhost:8081
    description: Local development server
tags:
  - name: Authentication
    description: Registration, login, logout, and shipping profile APIs
  - name: Products
    description: Product browsing APIs
  - name: Orders
    description: Authenticated order APIs
paths:
  /api/auth/register:
    post:
      tags:
        - Authentication
      summary: Register a new user and start a session
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
            example:
              username: Carlos
              account: carlos001
              password: "123456"
      responses:
        "200":
          description: Registered user information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        "400":
          description: Invalid request body
  /api/auth/login:
    post:
      tags:
        - Authentication
      summary: Login with account and password
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            example:
              account: carlos001
              password: "123456"
      responses:
        "200":
          description: Logged-in user information
          headers:
            Set-Cookie:
              description: Session cookie returned by Spring Boot
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        "400":
          description: Invalid account or password
  /api/auth/me:
    get:
      tags:
        - Authentication
      summary: Get the current session user
      operationId: getCurrentUser
      security:
        - sessionCookie: []
      responses:
        "200":
          description: Current user, or null if not logged in
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
  /api/auth/profile:
    put:
      tags:
        - Authentication
      summary: Update the current user's shipping profile
      operationId: updateProfile
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProfileRequest'
            example:
              recipientName: Zhang San
              phone: "13800138000"
              address: Beijing Chaoyang District
      responses:
        "200":
          description: Updated user information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        "401":
          description: Login required
  /api/auth/logout:
    post:
      tags:
        - Authentication
      summary: Logout and invalidate the session
      operationId: logout
      security:
        - sessionCookie: []
      responses:
        "200":
          description: Logout result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        "401":
          description: Login required
  /api/products:
    get:
      tags:
        - Products
      summary: List all products
      operationId: listProducts
      responses:
        "200":
          description: Product list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
  /api/products/{id}:
    get:
      tags:
        - Products
      summary: Find one product by id
      operationId: getProductById
      parameters:
        - name: id
          in: path
          required: true
          description: Product id
          schema:
            type: integer
            format: int64
            example: 1
      responses:
        "200":
          description: Product detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        "400":
          description: Product does not exist
  /api/orders:
    get:
      tags:
        - Orders
      summary: List current user's orders
      operationId: listOrders
      security:
        - sessionCookie: []
      parameters:
        - name: X-Client-Meta-Tag
          in: header
          required: false
          description: Demo header copied from an HTML meta tag and checked by a filter
          schema:
            type: string
            example: allow-orders
      responses:
        "200":
          description: Current user's order list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OrderResponse'
        "401":
          description: Login required
        "403":
          description: Blocked by meta-tag behavior filter
    post:
      tags:
        - Orders
      summary: Create an order for the current user
      operationId: createOrder
      security:
        - sessionCookie: []
      parameters:
        - name: X-Client-Meta-Tag
          in: header
          required: false
          description: Demo header copied from an HTML meta tag and checked by a filter
          schema:
            type: string
            example: allow-orders
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
            example:
              items:
                - productId: 1
                  quantity: 2
      responses:
        "200":
          description: Created order
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        "401":
          description: Login required
        "403":
          description: Blocked by meta-tag behavior filter
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: JSESSIONID
  schemas:
    RegisterRequest:
      type: object
      required:
        - username
        - account
        - password
      properties:
        username:
          type: string
          example: Carlos
        account:
          type: string
          example: carlos001
        password:
          type: string
          minLength: 6
          example: "123456"
    LoginRequest:
      type: object
      required:
        - account
        - password
      properties:
        account:
          type: string
          example: carlos001
        password:
          type: string
          example: "123456"
    ProfileRequest:
      type: object
      required:
        - recipientName
        - phone
        - address
      properties:
        recipientName:
          type: string
          example: Zhang San
        phone:
          type: string
          example: "13800138000"
        address:
          type: string
          example: Beijing Chaoyang District
    CreateOrderRequest:
      type: object
      required:
        - items
      properties:
        items:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/CartItemRequest'
    CartItemRequest:
      type: object
      required:
        - productId
        - quantity
      properties:
        productId:
          type: integer
          format: int64
          example: 1
        quantity:
          type: integer
          minimum: 1
          example: 2
    UserResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        username:
          type: string
          example: Carlos
        account:
          type: string
          example: carlos001
        recipientName:
          type: string
          nullable: true
          example: Zhang San
        phone:
          type: string
          nullable: true
          example: "13800138000"
        address:
          type: string
          nullable: true
          example: Beijing Chaoyang District
        profileComplete:
          type: boolean
          example: true
    Product:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        name:
          type: string
          example: Wireless Mouse
        description:
          type: string
          example: Ergonomic wireless mouse
        price:
          type: number
          format: decimal
          example: 59.90
        imageUrl:
          type: string
          example: https://images.unsplash.com/photo-1527814050087-3793815479db
        stock:
          type: integer
          example: 100
    OrderResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        account:
          type: string
          example: carlos001
        customerName:
          type: string
          example: Zhang San
        phone:
          type: string
          example: "13800138000"
        address:
          type: string
          example: Beijing Chaoyang District
        totalAmount:
          type: number
          format: decimal
          example: 119.80
        createdAt:
          type: string
          format: date-time
          example: "2026-05-24T10:30:00"
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItemResponse'
    OrderItemResponse:
      type: object
      properties:
        productId:
          type: integer
          format: int64
          example: 1
        productName:
          type: string
          example: Wireless Mouse
        unitPrice:
          type: number
          format: decimal
          example: 59.90
        quantity:
          type: integer
          example: 2
        lineTotal:
          type: number
          format: decimal
          example: 119.80
    MessageResponse:
      type: object
      properties:
        message:
          type: string
          example: logout success